requirements-prod.txt
--extra-index-url https://pypi.tuna.tsinghua.edu.cn/simple

aiofiles==23.1.0 ; python_version >= "3.9" and python_version < "4.0"
aiomysql==0.2.0 ; python_version >= "3.9" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.9" and python_full_version <= "3.11.2"
babel==2.12.1 ; python_version >= "3.9" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.6 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
exceptiongroup==1.1.2 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.100.0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
httptools==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.9" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.9" and python_version < "4.0"
pendulum==2.1.2 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.11 ; python_version >= "3.9" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.9" and python_version < "4.0"
pymysql==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
pypika-tortoise==0.1.6 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.9" and python_version < "4.0"
pytzdata==2020.1 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
redis==4.6.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.9" and python_version < "4.0"
style==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
tortoise-orm==0.19.3 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.9" and python_version < "4.0"
update==0.0.1 ; python_version >= "3.9" and python_version < "4.0"
uvicorn[standard]==0.23.1 ; python_version >= "3.9" and python_version < "4.0"
uvloop==0.17.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.9" and python_version < "4.0"
watchfiles==0.19.0 ; python_version >= "3.9" and python_version < "4.0"
websockets==11.0.3 ; python_version >= "3.9" and python_version < "4.0"
